first work prototype

+++ app/src/main/java/main/src/etherscan/data/models/ETHModel.kt
@@ -1,7 +1,7 @@
 package main.src.etherscan.data.models
 
 data class ETHModel(
-    val balance: Int = 0,
-    val totalIn: Int = 0,
-    val totalOut: Int = 0
+    val balance: Double = 0.0,
+    val totalIn: Double = 0.0,
+    val totalOut: Double = 0.0
 )
